Refactor Streaming out of dataframes
This allows us to bring more collections onto the same framework
